move validateBackendPermissions to separate file to avoid circular dependency error
